Refactoring and cleaning up acceptance tests
